Add checksum functionality to the plugin
With this, when XRootD needs to query for a checksum, we make the
request to the origin (as appropriate) using a HEAD call and return
the checksum back to the client.

Since the objects are assumed to be immutable, the checksum is kept
in memory for a modest amount of time, allowing repeated use of the
object to not cause load on the origin.

Includes unit tests.